+++ nequip/train/callbacks/__init__.py
@@ -7,12 +7,14 @@ from .loss_coeff_scheduler import (
 from .loss_coeff_monitor import LossCoefficientMonitor
 from .write_xyz import TestTimeXYZFileWriter
 from .wandb_watch import WandbWatch
+from .tf32_scheduler import TF32Scheduler
 
 __all__ = [
     SoftAdapt,
     LossCoefficientScheduler,
     LinearLossCoefficientScheduler,
     LossCoefficientMonitor,
+    TF32Scheduler,
     TestTimeXYZFileWriter,
     WandbWatch,
 ]